ES6 syntax

+++ gulpfile.js
@@ -1,12 +1,12 @@
-var gulp = require('gulp');
-var concat = require('gulp-concat');
-var sourcemaps = require('gulp-sourcemaps');
-var header = require('gulp-header');
-var rename = require('gulp-rename');
-var fs = require('fs');
-var notifier = require('node-notifier');
+const gulp = require('gulp');
+const concat = require('gulp-concat');
+const sourcemaps = require('gulp-sourcemaps');
+const header = require('gulp-header');
+const rename = require('gulp-rename');
+const fs = require('fs');
+const notifier = require('node-notifier');
 
-var compile, uglify, babel, fast;
+let compile, uglify, babel, fast;
 gulp.task('enableCompile', function(done){
   uglify = require('gulp-uglify-es').default;
   babel = require("gulp-babel");
@@ -15,9 +15,9 @@ gulp.task('enableCompile', function(done){
   done && done();
 });
 
-var dateMinify = {};
+const dateMinify = {};
 // var dateMinify = {mapFile:function(path){return path.replace('js.map', Date.now()+'.js.map')}};
-var theHeader = `/*
+const theHeader = `/*
   ScarletsFrame
   A frontend library for Scarlets Framework that support
   lazy page load and element binding that can help
@@ -26,7 +26,7 @@ var theHeader = `/*
   https://github.com/ScarletsFiction/ScarletsFrame
 */\n`;
 
-var theOrderES5 = [
+const theOrderES5 = [
   'src/sf-polyfill.js',
   'src/sf-a_init.js',
   'src/sf-loader.js',
@@ -40,8 +40,8 @@ var theOrderES5 = [
   'src/sf-z_end.js',
 ];
 
-var theOrderES6 = theOrderES5.slice(1);
-var devTest = theOrderES6.slice(0);
+const theOrderES6 = theOrderES5.slice(1);
+const devTest = theOrderES6.slice(0);
 devTest.splice(-2, 1); // Remove '!src/sf-hot-reload.js'
 
 gulp.task('watch-js', function(){
@@ -61,8 +61,8 @@ gulp.task('watch-js', function(){
     });
 });
 
-gulp.task('js-es5', function(){
-  return gulp.src(theOrderES5)
+gulp.task('js-es5', ()=>
+  gulp.src(theOrderES5)
     .pipe(sourcemaps.init())
     .pipe(concat('scarletsframe.js'))
     .pipe(gulp.dest('dist'))
@@ -75,7 +75,8 @@ gulp.task('js-es5', function(){
               ie: "11"
             },
             modules: false,
-            loose: false
+            loose: false,
+            shippedProposals: true
           }
         ]
       ]
@@ -84,28 +85,28 @@ gulp.task('js-es5', function(){
     .pipe(header(theHeader))
     .pipe(rename({extname:'.min.js'}))
     .pipe(sourcemaps.write('.', dateMinify))
-    .pipe(gulp.dest('dist'));
-});
+    .pipe(gulp.dest('dist'))
+);
 
-gulp.task('js-es6', function(){
-  return gulp.src(theOrderES6)
+gulp.task('js-es6', ()=>
+  gulp.src(theOrderES6)
     .pipe(sourcemaps.init())
     .pipe(concat('scarletsframe.es6.js'))
     .pipe(uglify())
     .pipe(header(theHeader))
     .pipe(sourcemaps.write('.', dateMinify))
-    .pipe(gulp.dest('dist'));
-});
+    .pipe(gulp.dest('dist'))
+);
 
-gulp.task('js-hot', function(){
-  return gulp.src(devTest)
+gulp.task('js-hot', ()=>
+  gulp.src(devTest)
     .pipe(sourcemaps.init())
     .pipe(concat('scarletsframe.hot.js'))
     .pipe(uglify())
     .pipe(header(theHeader))
     .pipe(sourcemaps.write('.', dateMinify))
-    .pipe(gulp.dest('dist'));
-});
+    .pipe(gulp.dest('dist'))
+);
 
 gulp.task('default', function(){
   require('./tests/server.js');
@@ -129,10 +130,12 @@ function swallowError(error){
 
 function removeOldMap(path){
   fs.readdir(path, function(err, files){
-     for (var i = 0, len = files.length; i < len; i++) {
-        if(files[i].match(/.*\.(js|css)\.map/) !== null){
-          try{fs.unlinkSync(path+files[i]);}catch(e){}
-        }
+     if (files) {
+       for (let i = 0, len = files.length; i < len; i++) {
+          if(files[i].match(/.*\.(js|css)\.map/) !== null){
+            try{fs.unlinkSync(path+files[i]);}catch(e){}
+          }
+       }
      }
   });
 }
